Remove old sysinfo and finish sys
